metrics-migrate-stats/requirements.txt
psycopg2-binary==2.8.6
geoip2
